# docs/v3/openapi3.json
{
	"components": {
		"schemas": {
...
			"datastore.ApiKey": {
				"properties": {
					"header_name": {
						"type": "string"
					},
					"header_value": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"datastore.EndpointAuthentication": {
				"properties": {
					"api_key": {
						"$ref": "#/components/schemas/datastore.ApiKey"
					},
					"type": {
						"$ref": "#/components/schemas/datastore.EndpointAuthenticationType"
					}
				},
				"type": "object"
			},
			"datastore.EndpointAuthenticationType": {
				"enum": [
					"api_key"
				],
				"type": "string",
				"x-enum-varnames": [
					"APIKeyAuthentication"
				]
			},
			"datastore.EndpointStatus": {
				"enum": [
					"active",
					"inactive",
					"pending",
					"paused"
				],
				"type": "string",
				"x-enum-varnames": [
					"ActiveEndpointStatus",
					"InactiveEndpointStatus",
					"PendingEndpointStatus",
					"PausedEndpointStatus"
				]
			},
...
			"datastore.PaginationData": {
				"properties": {
					"has_next_page": {
						"type": "boolean"
					},
					"has_prev_page": {
						"type": "boolean"
					},
					"next_page_cursor": {
						"type": "string"
					},
					"per_page": {
						"type": "integer"
					},
					"prev_page_cursor": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"datastore.Secret": {
				"properties": {
					"created_at": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string"
					},
					"expires_at": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"handlers.Stub": {
				"type": "object"
			},
...
			"models.ApiKey": {
				"properties": {
					"header_name": {
						"type": "string"
					},
					"header_value": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"models.CreateEndpoint": {
				"properties": {
					"advanced_signatures": {
						"description": "Convoy supports two [signature formats](https://getconvoy.io/docs/manual/signatures)\n-- simple or advanced. If left unspecified, we default to false.",
						"type": "boolean"
					},
					"appID": {
						"description": "Deprecated but necessary for backward compatibility",
						"type": "string"
					},
					"authentication": {
						"allOf": [
							{
								"$ref": "#/components/schemas/models.EndpointAuthentication"
							}
						],
						"description": "This is used to define any custom authentication required by the endpoint. This\nshouldn't be needed often because webhook endpoints usually should be exposed to\nthe internet."
					},
					"description": {
						"description": "Human-readable description of the endpoint. Think of this as metadata describing\nthe endpoint",
						"type": "string"
					},
					"http_timeout": {
						"description": "Define endpoint http timeout in seconds.",
						"type": "integer"
					},
					"is_disabled": {
						"description": "This is used to manually enable/disable the endpoint.",
						"type": "boolean"
					},
					"name": {
						"description": "Endpoint name.",
						"type": "string"
					},
					"owner_id": {
						"description": "The OwnerID is used to group more than one endpoint together to achieve\n[fanout](https://getconvoy.io/docs/manual/endpoints#Endpoint%20Owner%20ID)",
						"type": "string"
					},
					"rate_limit": {
						"description": "Rate limit is the total number of requests to be sent to an endpoint in\nthe time duration specified in RateLimitDuration",
						"type": "integer"
					},
					"rate_limit_duration": {
						"description": "Rate limit duration specifies the time range for the rate limit.",
						"type": "integer"
					},
					"secret": {
						"description": "Endpoint's webhook secret. If not provided, Convoy autogenerates one for the endpoint.",
						"type": "string"
					},
					"slack_webhook_url": {
						"description": "Slack webhook URL is an alternative method to support email where endpoint developers\ncan receive failure notifications on a slack channel.",
						"type": "string"
					},
					"support_email": {
						"description": "Endpoint developers support email. This is used for communicating endpoint state\nchanges. You should always turn this on when disabling endpoints are enabled.",
						"type": "string"
					},
					"url": {
						"description": "URL is the endpoint's URL prefixed with https. non-https urls are currently\nnot supported.",
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"models.EndpointAuthentication": {
				"properties": {
					"api_key": {
						"$ref": "#/components/schemas/models.ApiKey"
					},
					"type": {
						"$ref": "#/components/schemas/datastore.EndpointAuthenticationType"
					}
				},
				"type": "object"
			},
			"models.EndpointResponse": {
				"properties": {
					"advanced_signatures": {
						"type": "boolean"
					},
					"authentication": {
						"$ref": "#/components/schemas/datastore.EndpointAuthentication"
					},
					"created_at": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"events": {
						"type": "integer"
					},
					"http_timeout": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					},
					"owner_id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"rate_limit": {
						"type": "integer"
					},
					"rate_limit_duration": {
						"type": "integer"
					},
					"secrets": {
						"items": {
							"$ref": "#/components/schemas/datastore.Secret"
						},
						"type": "array"
					},
					"slack_webhook_url": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/datastore.EndpointStatus"
					},
					"support_email": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"models.PagedResponse": {
				"properties": {
					"content": {},
					"pagination": {
						"$ref": "#/components/schemas/datastore.PaginationData"
					}
				},
				"type": "object"
			},
...
			"util.ServerResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "support@getconvoy.io",
			"name": "Convoy Support",
			"url": "https://getconvoy.io/docs"
		},
		"description": "Convoy is a fast and secure webhooks proxy. This document contains datastore.s API specification.",
		"license": {
			"name": "Mozilla Public License 2.0",
			"url": "https://www.mozilla.org/en-US/MPL/2.0/"
		},
		"termsOfService": "https://getconvoy.io/terms",
		"title": "Convoy API Reference",
		"version": "24.1.4"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/v1/projects/{projectID}/endpoints": {
			"get": {
				"description": "This endpoint fetches an endpoints",
				"operationId": "GetEndpoints",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"schema": {
							"enum": [
								"next",
								"prev"
							],
							"type": "string"
						},
						"x-enum-varnames": [
							"Next",
							"Prev"
						]
					},
					{
						"description": "A pagination cursor to fetch the next page of a list",
						"in": "query",
						"name": "next_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The owner ID of the endpoint",
						"in": "query",
						"name": "ownerId",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The number of items to return per page",
						"in": "query",
						"name": "perPage",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "A pagination cursor to fetch the previous page of a list",
						"in": "query",
						"name": "prev_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The name of the endpoint",
						"in": "query",
						"name": "q",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order, values are `ASC` or `DESC`, defaults to `DESC`",
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/models.PagedResponse"
														},
														{
															"properties": {
																"content": {
																	"items": {
																		"$ref": "#/components/schemas/models.EndpointResponse"
																	},
																	"type": "array"
																}
															},
															"type": "object"
														}
													]
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "List all endpoints",
				"tags": [
					"Endpoints"
				]
			},
			"post": {
				"description": "This endpoint creates an endpoint",
				"operationId": "CreateEndpoint",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.CreateEndpoint"
							}
						}
					},
					"description": "Endpoint Details",
					"required": true,
					"x-originalParamName": "endpoint"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.EndpointResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Create an endpoint",
				"tags": [
					"Endpoints"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://dashboard.getconvoy.io/api"
		}
	],
...
}